uptadet TitleEntity(added  list of seasons), added season entity, dto, repo, service, get method in seasonController

+++ backend/demo/src/main/java/project_z/demo/repositories/TitleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import project_z.demo.entity.TitleEntity;
 
 @Repository
-public interface  TitleRepository extends CrudRepository<TitleEntity, Integer>{
+public interface  TitleRepository extends CrudRepository<TitleEntity, Long>{
  // @Query("SELECT t FROM TitleEntity t WHERE t.user.id = :userId")
    // List<TitleEntity> findByUserId(@Param("userId") UUID userId);
 }
