+++ backend/demo/src/main/java/project_z/demo/services/TitleService.java
@@ -9,10 +9,10 @@ import project_z.demo.entity.TitleEntity;
 public interface  TitleService {
     TitleEntity createTitle(TitleEntity title );
     List<TitleEntity> findAll();
-    Optional<TitleEntity> findOne(Integer titleId);
-    boolean isExists (Integer titleId);
-    TitleEntity partialUpdate(Integer titleId, TitleEntity titleEntity);
-    void deleteById(Integer id);
+    Optional<TitleEntity> findOne(Long titleId);
+    boolean isExists (Long titleId);
+    TitleEntity partialUpdate(Long titleId, TitleEntity titleEntity);
+    void deleteById(Long id);
     List<TitleEntity> addTitle(TitleEntity titleEntity,  UUID userId);
     List<TitleEntity> getWatchedList(UUID userid);
     List<TitleEntity> getWatchList(UUID userid);
